+++ lib/active_record_proxy_adapters/configuration.rb
@@ -1,32 +1,17 @@
 # frozen_string_literal: true
 
-require "active_support/core_ext/integer/time"
-require "active_record_proxy_adapters/synchronizable_configuration"
 require "active_record_proxy_adapters/cache_configuration"
 require "active_record_proxy_adapters/context"
+require "active_record_proxy_adapters/database_configuration"
+require "active_record_proxy_adapters/synchronizable_configuration"
+require "active_support/core_ext/integer/time"
 
 module ActiveRecordProxyAdapters
   # Provides a global configuration object to configure how the proxy should behave.
   class Configuration
     include SynchronizableConfiguration
 
-    PROXY_DELAY                   = 2.seconds.freeze
-    CHECKOUT_TIMEOUT              = 2.seconds.freeze
-    LOG_SUBSCRIBER_PRIMARY_PREFIX = proc { |event| "#{event.payload[:connection].class::ADAPTER_NAME} Primary" }.freeze
-    LOG_SUBSCRIBER_REPLICA_PREFIX = proc { |event| "#{event.payload[:connection].class::ADAPTER_NAME} Replica" }.freeze
-
-    # @return [ActiveSupport::Duration] How long the proxy should reroute all read requests to the primary database
-    #   since the latest write. Defaults to PROXY_DELAY. Thread safe.
-    attr_reader :proxy_delay
-    # @return [ActiveSupport::Duration] How long the proxy should wait for a connection from the replica pool.
-    #   Defaults to CHECKOUT_TIMEOUT. Thread safe.
-    attr_reader :checkout_timeout
-
-    # @return [Proc] Prefix for the log subscriber when the primary database is used. Thread safe.
-    attr_reader :log_subscriber_primary_prefix
-
-    # @return [Proc] Prefix for the log subscriber when the replica database is used. Thread safe.
-    attr_reader :log_subscriber_replica_prefix
+    DEFAULT_DATABASE_NAME = :primary
 
     # @return [Class] The context that is used to store the current request's state.
     attr_reader :context_store
@@ -34,46 +19,48 @@ module ActiveRecordProxyAdapters
     def initialize
       @lock = Monitor.new
 
-      self.proxy_delay                   = PROXY_DELAY
-      self.checkout_timeout              = CHECKOUT_TIMEOUT
-      self.log_subscriber_primary_prefix = LOG_SUBSCRIBER_PRIMARY_PREFIX
-      self.log_subscriber_replica_prefix = LOG_SUBSCRIBER_REPLICA_PREFIX
-      self.cache_configuration           = CacheConfiguration.new(@lock)
-      self.context_store                 = ActiveRecordProxyAdapters::Context
+      self.cache_configuration = CacheConfiguration.new(@lock)
+      self.context_store       = ActiveRecordProxyAdapters::Context
+      @database_configurations = {}
     end
 
     def log_subscriber_primary_prefix=(prefix)
-      prefix_proc = prefix.is_a?(Proc) ? prefix : proc { prefix.to_s }
+      default_database_config.log_subscriber_primary_prefix = prefix
+    end
 
-      synchronize_update(:log_subscriber_primary_prefix, from: @log_subscriber_primary_prefix, to: prefix_proc) do
-        @log_subscriber_primary_prefix = prefix_proc
-      end
+    def log_subscriber_primary_prefix
+      default_database_config.log_subscriber_primary_prefix
     end
 
     def log_subscriber_replica_prefix=(prefix)
-      prefix_proc = prefix.is_a?(Proc) ? prefix : proc { prefix.to_s }
+      default_database_config.log_subscriber_replica_prefix = prefix
+    end
 
-      synchronize_update(:log_subscriber_replica_prefix, from: @log_subscriber_replica_prefix, to: prefix_proc) do
-        @log_subscriber_replica_prefix = prefix_proc
-      end
+    def log_subscriber_replica_prefix
+      default_database_config.log_subscriber_replica_prefix
+    end
+
+    def proxy_delay
+      default_database_config.proxy_delay
     end
 
     def proxy_delay=(proxy_delay)
-      synchronize_update(:proxy_delay, from: @proxy_delay, to: proxy_delay) do
-        @proxy_delay = proxy_delay
-      end
+      default_database_config.proxy_delay = proxy_delay
+    end
+
+    def checkout_timeout
+      default_database_config.checkout_timeout
     end
 
     def checkout_timeout=(checkout_timeout)
-      synchronize_update(:checkout_timeout, from: @checkout_timeout, to: checkout_timeout) do
-        @checkout_timeout = checkout_timeout
-      end
+      default_database_config.checkout_timeout = checkout_timeout
     end
 
-    def context_store=(context_store)
-      synchronize_update(:context_store, from: @context_store, to: context_store) do
-        @context_store = context_store
-      end
+    def database(database_name)
+      key = database_name.to_s
+      lock.synchronize { @database_configurations[key] ||= DatabaseConfiguration.new }
+
+      block_given? ? yield(database_configurations[key]) : database_configurations[key]
     end
 
     def cache
@@ -82,12 +69,21 @@ module ActiveRecordProxyAdapters
 
     private
 
-    # @return [CacheConfiguration] The cache configuration for the proxy adapters.
-    attr_reader :cache_configuration
+    attr_reader :cache_configuration, :database_configurations, :lock
+
+    def default_database_config
+      database(DEFAULT_DATABASE_NAME)
+    end
 
     def cache_configuration=(cache_configuration)
       synchronize_update(:cache_configuration, from: @cache_configuration, to: cache_configuration) do
         @cache_configuration = cache_configuration
+      end
+    end
+
+    def context_store=(context_store)
+      synchronize_update(:context_store, from: @context_store, to: context_store) do
+        @context_store = context_store
       end
     end
   end