+++ lib/active_record_proxy_adapters/log_subscriber.rb
@@ -2,6 +2,8 @@
 
 module ActiveRecordProxyAdapters
   class LogSubscriber < ActiveRecord::LogSubscriber # rubocop:disable Style/Documentation
+    include Mixin::Configuration
+
     attach_to :active_record
 
     IGNORE_PAYLOAD_NAMES = %w[SCHEMA EXPLAIN].freeze
@@ -19,23 +21,27 @@ module ActiveRecordProxyAdapters
     protected
 
     def database_instance_prefix_for(event)
-      connection = event.payload[:connection]
-      config = connection.instance_variable_get(:@config)
-      prefix = if config[:replica] || config["replica"]
-                 log_subscriber_replica_prefix
+      connection      = event.payload[:connection]
+      db_config       = connection.pool.try(:db_config) || NullConfig.new # AR 7.0.x does not support "NullConfig"
+      connection_name = db_config.name
+
+      prefix = if db_config.replica?
+                 log_subscriber_replica_prefix(connection_name)
                else
-                 log_subscriber_primary_prefix
+                 log_subscriber_primary_prefix(connection_name)
                end
 
       "[#{prefix.call(event)}]"
     end
 
-    private
+    class NullConfig # rubocop:disable Style/Documentation
+      def method_missing(...)
+        nil
+      end
 
-    delegate :log_subscriber_primary_prefix, :log_subscriber_replica_prefix, to: :config
-
-    def config
-      ActiveRecordProxyAdapters.config
+      def respond_to_missing?(*)
+        true
+      end
     end
   end
 end