+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add support for configuring a Context API
 - Add Rack middleware to store context in session
+- Allow separate configuration blocks per database name
 
 ## [0.6.2, 0.5.2, 0.4.8] - 2025-07-21